drivers: audio i2s: Move audio i2s API header file to audio folder
- to align with documentation, move the i2s header to audio folder
- align CMakeLists.txt, Kconfig, prj.conf
