+++ drivers/audio/i2s/CMakeLists.txt
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 
-zephyr_syscall_header(${ZEPHYR_BASE}/include/zephyr/drivers/i2s.h)
+zephyr_syscall_header(${ZEPHYR_BASE}/include/zephyr/audio/i2s.h)
 
 zephyr_library()
