+++ drivers/audio/i2s/i2s_max32.c
@@ -5,7 +5,7 @@
 
 #define DT_DRV_COMPAT adi_max32_i2s
 
-#include <zephyr/drivers/i2s.h>
+#include <zephyr/audio/i2s.h>
 #include <zephyr/drivers/dma.h>
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/device.h>
