+++ drivers/audio/i2s/i2s_mcux_sai.c
@@ -18,7 +18,7 @@
 #include <zephyr/device.h>
 #include <zephyr/init.h>
 #include <zephyr/drivers/dma.h>
-#include <zephyr/drivers/i2s.h>
+#include <zephyr/audio/i2s.h>
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/drivers/clock_control.h>
 #ifdef CONFIG_CLOCK_CONTROL_MCUX_CCM
