+++ drivers/audio/i2s/i2s_nrf_tdm.c
@@ -11,7 +11,7 @@
 #include <haly/nrfy_gpio.h>
 #include <zephyr/drivers/clock_control/nrf_clock_control.h>
 #include <zephyr/dt-bindings/clock/nrf-auxpll.h>
-#include <zephyr/drivers/i2s.h>
+#include <zephyr/audio/i2s.h>
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/irq.h>
 #include <zephyr/kernel.h>
