+++ drivers/audio/i2s/i2s_silabs_siwx91x.c
@@ -11,7 +11,7 @@
 #include <zephyr/sys/__assert.h>
 #include <zephyr/drivers/dma.h>
 #include <zephyr/drivers/pinctrl.h>
-#include <zephyr/drivers/i2s.h>
+#include <zephyr/audio/i2s.h>
 #include <zephyr/drivers/clock_control.h>
 #include <zephyr/irq.h>
 #include <zephyr/sys/util.h>
