+++ drivers/audio/i2s/i2s_stm32_sai.c
@@ -9,7 +9,7 @@
 
 #include <string.h>
 #include <zephyr/drivers/dma.h>
-#include <zephyr/drivers/i2s.h>
+#include <zephyr/audio/i2s.h>
 #include <zephyr/drivers/dma/dma_stm32.h>
 #include <soc.h>
 #include <zephyr/drivers/clock_control/stm32_clock_control.h>
