+++ include/zephyr/audio/codec.h
@@ -25,7 +25,7 @@
  */
 
 #include <errno.h>
-#include <zephyr/drivers/i2s.h>
+#include <zephyr/audio/i2s.h>
 
 #include <zephyr/audio/audio_caps.h> /* Include common audio caps */
