+++ samples/drivers/i2s/output/src/main.c
@@ -6,7 +6,7 @@
 
 #include <stdio.h>
 #include <zephyr/kernel.h>
-#include <zephyr/drivers/i2s.h>
+#include <zephyr/audio/i2s.h>
 #include <zephyr/sys/iterable_sections.h>
 
 #define SAMPLE_NO 64
