+++ tests/drivers/i2s/i2s_api/src/test_i2s_loopback.c
@@ -7,7 +7,7 @@
 
 #include <zephyr/kernel.h>
 #include <zephyr/ztest.h>
-#include <zephyr/drivers/i2s.h>
+#include <zephyr/audio/i2s.h>
 #include "i2s_api_test.h"
 
 /** @brief Short I2S transfer.
