+++ tests/drivers/i2s/i2s_api/src/test_i2s_states.c
@@ -6,7 +6,7 @@
 
 #include <zephyr/kernel.h>
 #include <zephyr/ztest.h>
-#include <zephyr/drivers/i2s.h>
+#include <zephyr/audio/i2s.h>
 #include "i2s_api_test.h"
 
 /** @brief Verify all failure cases in NOT_READY state.
